+++ packages/projectlint/src/codemod/transforms/__fixtures__/lint-config-to-applint-spec/iceworks-spec-config/.eslintrc.js
@@ -1,0 +1,14 @@
+const { getESLintConfig } = require('@iceworks/spec');
+const config = getESLintConfig('react-ts');
+
+module.exports = {
+  ...config,
+  overrides: [
+    {
+      files: ['**/src/apis/**'],
+      rules: {
+        'react-hooks/rules-of-hooks': 0,
+      },
+    },
+  ],
+};